+++ ProgramFunctions/Theme.fnc.php
@@ -48,8 +48,7 @@ function ThemeLiveUpdate( $new_theme, $old_theme, $default = true )
 			AND PROGRAM='Preferences'
 			AND TITLE='THEME'" );
 
-		if ( ! $real_user_preferred_theme
-			|| $real_user_preferred_theme === $new_theme )
+		if ( $real_user_preferred_theme === $new_theme )
 		{
 			// User already had that Theme loaded, do nothing.
 			return false;