+++ modules/School_Setup/Configuration.php
@@ -73,6 +73,8 @@ else
 		{
 			$updated = $numeric_error = false;
 
+			$old_theme = Config( 'THEME' );
+
 			$_REQUEST['values']['CONFIG'] = issetVal( $_REQUEST['values']['CONFIG'] );
 
 			foreach ( (array) $_REQUEST['values']['CONFIG'] as $column => $value )
@@ -136,8 +138,6 @@ else
 			{
 				$error[] = _( 'Please enter valid Numeric data.' );
 			}
-
-			$old_theme = Config( 'THEME' );
 
 			// Theme changed? Update it live!
 			ThemeLiveUpdate( Config( 'THEME' ), $old_theme );
@@ -647,60 +647,4 @@ else
 
 		echo '</form>';
 	}
-}
-
-/**
- * Theme live update.
- * Configured theme has changed? Update it live!
- * Updates the stylesheet.css file to the new theme directory,
- * Using a Javascript snippet.
- *
- * @todo use it Preferences too!
- *
- * Local function
- *
- * @since  3.0
- *
- * @param  string  $new_theme New theme name / directory.
- * @param  string  $old_theme Old theme name / directory.
- * @param  boolean $default   Is default theme (Configuration.php) or Preferred theme (Preferences.php)?
- * @return boolean False if has not changed, else true.
- */
-function _themeLiveUpdate( $new_theme, $old_theme, $default = true )
-{
-	if ( ! $new_theme
-		|| ! $old_theme
-		|| $new_theme === $old_theme )
-	{
-		return false;
-	}
-
-	if ( ! $default
-		&& Config( 'THEME_FORCE' ) )
-	{
-		// Theme forced, we should not be able to change it anyway!
-		return false;
-	}
-
-	// If not Forcing theme, update admin Preferred theme too.
-
-	if ( $default
-		&& ! Config( 'THEME_FORCE' )
-		&& Preferences( 'THEME' ) !== $new_theme )
-	{
-		// TODO.
-	}
-
-	// Update stylesheet(s) href. ?>
-	<script>
-	$('link[href^="assets/themes"]').each(function(){
-		$(this).attr('href', $(this).attr('href').replace(
-			<?php echo json_encode( $old_theme ); ?>,
-			<?php echo json_encode( $new_theme ); ?>
-		) );
-	});
-	</script>
-	<?php
-
-	return true;
 }