+++ CHANGES.md
@@ -9,6 +9,7 @@ Changes in 6.2
 - HTML Seats, Credit Hours & Credits: use "number" input type in Courses.php & Courses.fnc.php
 - HTML Attendance & Food Service tabs: use "number" input type in Configuration.php
 - HTML use Tab instead of fieldset in Configuration.php
+- Fix Theme live update in Configuration.php & Theme.fnc.php
 
 Changes in 6.1
 --------------